refactor: lift cross-chain-swap form state into shared XSwap providers
Extract URL/path-driven form layer (chainId0/chainId1, token params,
swapAmount, tradeId, switchTokens, URL sync) from the existing LiFi
provider into a new XSwapFormProvider. The LiFi provider now consumes
the form context and owns only LiFi-specific concerns (selectedBridge,
routeOrder, Token resolution, recipient).

Renames adopt the codebase's "XSwap" vocabulary:
- DerivedstateCrossChainSwapProvider → LifiXSwapProvider
- useDerivedStateCrossChainSwap → useLifiXSwap
- useCrossChainTradeRoutes (LiFi wrapper) → useLifiXSwapTradeRoutes
- useSelectedCrossChainTradeRoute → useLifiXSwapSelectedTradeRoute
- New CrossChainSwapFormProvider → XSwapFormProvider
- New useCrossChainSwapFormState → useXSwapForm
- File: derivedstate-cross-chain-swap-provider.tsx → lifi-xswap-provider.tsx
- File: cross-chain-swap-form-provider.tsx (new) → xswap-form-provider.tsx

The underlying react-query hook useCrossChainTradeRoutes is unchanged.
Public hook return shapes (state/mutate/isLoading) are preserved so
widget consumers only update their imports.
