+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/_ui/cross-chain-swap-trade-review-dialog/hooks/use-cross-chain-swap-trade-review-pre.ts
@@ -4,14 +4,14 @@ import { useRef, useState } from 'react'
 import type { XSwapSupportedChainId } from 'src/config'
 import { useSlippageTolerance } from 'src/lib/hooks/useSlippageTolerance'
 import { StepState } from '../../cross-chain-swap-confirmation-dialog'
-import type { UseSelectedCrossChainTradeRouteReturn } from '../../derivedstate-cross-chain-swap-provider'
+import type { UseLifiXSwapSelectedTradeRouteReturn } from '../../lifi-xswap-provider'
 
 export function useCrossChainSwapTradeReviewPre<
   TChainId0 extends XSwapSupportedChainId,
   TChainId1 extends XSwapSupportedChainId,
 >() {
   const [slippagePercent] = useSlippageTolerance()
-  const routeRef = useRef<UseSelectedCrossChainTradeRouteReturn<
+  const routeRef = useRef<UseLifiXSwapSelectedTradeRouteReturn<
     TChainId0,
     TChainId1
   > | null>(null)
