+++ apps/web/src/app/(networks)/(evm)/[chainId]/(trade)/cross-chain-swap/providers.tsx
@@ -2,7 +2,8 @@ import { getCrossChainSwapEdgeConfig } from './get-cross-chain-swap-edge-config'
 
 import type { XSwapSupportedChainId } from 'src/config'
 import { EdgeProvider } from 'src/providers/edge-config-provider'
-import { DerivedstateCrossChainSwapProvider } from './_ui/derivedstate-cross-chain-swap-provider'
+import { XSwapFormProvider } from './_ui/xswap-form-provider'
+import { LifiXSwapProvider } from './_ui/lifi-xswap-provider'
 
 export async function Providers({
   children,
@@ -12,9 +13,11 @@ export async function Providers({
 
   return (
     <EdgeProvider config={config}>
-      <DerivedstateCrossChainSwapProvider defaultChainId={chainId}>
-        {children}
-      </DerivedstateCrossChainSwapProvider>
+      <XSwapFormProvider defaultChainId={chainId}>
+        <LifiXSwapProvider>
+          {children}
+        </LifiXSwapProvider>
+      </XSwapFormProvider>
     </EdgeProvider>
   )
 }
